feat: splits views
