+++ apps/core/views/__init__.py
@@ -1,0 +1,7 @@
+from apps.core.views.base_view import BaseView  # noqa: F401
+from apps.core.views.maintenance_view import MaintenanceView  # noqa: F401
+from apps.core.views.manifest_view import ManifestView  # noqa: F401
+from apps.core.views.offline_view import OfflineView  # noqa: F401
+from apps.core.views.service_worker_view import ServiceWorkerView  # noqa: F401
+from apps.core.views.toast_htmx_view import ToastHTMXView  # noqa: F401
+from apps.core.views.welcome_partial_view import WelcomePartialView  # noqa: F401
