+++ apps/core/views/mixins/__init__.py
@@ -1,0 +1,1 @@
+from apps.core.views.mixins.maintenance_offline_mixin import MaintenanceOrOfflineView  # noqa: F401
