+++ apps/core/views/offline_view.py
@@ -1,0 +1,7 @@
+from django.views import generic
+
+from apps.core.views import mixins
+
+
+class OfflineView(mixins.MaintenanceOrOfflineView, generic.TemplateView):
+    pass
