+++ apps/core/views/service_worker_view.py
@@ -1,0 +1,6 @@
+from django.views import generic
+
+
+class ServiceWorkerView(generic.TemplateView):
+    template_name = "core/pwa/serviceworker.js"
+    content_type = "application/x-javascript"
